# swagger.yml
swagger: "2.0"
info:
  description: "Codecademy e-commerce REST API"
  version: "1.0.0"
  title: "E-commerce REST API"
schemes: 
  - http
host: localhost:3000
basePath: "/"
paths:
...
  /users/{userId}:
    get:
      summary: "Get user record"
      description: "Get user record"
      produces:
        - "application/json"
      parameters:
        - name: userId
          in: path
          required: true
          description: Numeric ID of the user to get
          schema:
            type: integer
      responses:
        200:
          description: 200 Success
          schema:
            type: object
            items:
        500:
          description: 500 Internal Server Error
          schema:
            type: object
            items:
  /users:
    put:
      summary: "Update user record"
      description: "Update user record (requires logged in user)"
      produces:
        - "application/json"
      parameters:
        - name: user
          in: body
          description: User login details to be updated
          required: true
          schema:
            type: object
            properties:
              email:
                type: string
              password: 
                type: string
              first_name:
                type: string
              last_name:
                type: string
      responses:
        200:
          description: 200 Success
          schema:
            type: object
            items:
        500:
          description: 500 Internal Server Error
          schema:
            type: object
            items:
...